requirements.txt
PySide6>=6.6.0
pystray>=0.19.4
Pillow>=10.0.0
requests>=2.31.0
schedule>=1.2.0
plyer>=2.1.0
pydantic>=2.5.0
APScheduler>=3.10.4
